pdf-ocr-pipeline/requirements.txt
google-generativeai
openai
anthropic
PyPDF2
streamlit
python-dotenv
